Brush smoothing and image transformation
Brushes now have draw points interpolated, creating a smooth curve. Brush spraying and image transformations are now available. GUI updated to allow for manipulation of new features. Remain feature yet to be done is the Radial Brush, however to be able to fully test it, the appropiate GUI work must be done.