Vectoring completed along with most layering.
Vectoring being fully completed, and the basics of layers being done as well. Layer features need to be integrated with framing, which needs it's features expanded upon. Filtering needs to be re-implemented, along with layer compiling. Minor additions need to be made to the raster mode, and the UI fully completed. This is the basic target functionality of the program.
Documentation will come later.
The first bug is aliasing on the top and bottom of the screen with vectors, and when filtering, on the sides as well.
The second known bug is occasionally the bounds for vectors are incorrectly calculated resulting in off-screen draws and uncentered rotations. This rarely occurs.